checkpoint

+++ install.sh
@@ -29,16 +29,21 @@ echo "Package list updated."
 
 # Install required packages
 echo "Installing required packages..."
-apt-get install -y python3 python3-pip git python3-pyqt5 pi-bluetooth python3-bluez || { echo "Package installation failed! Aborting..."; exit 1; }
+apt-get install -y python3 python3-pip git python3-pyqt5 pi-bluetooth python3-bluez build-essential cmake pkg-config libjpeg-dev libtiff5-dev libjasper-dev libpng-dev libavcodec-dev libavformat-dev libswscale-dev libv4l-dev libxvidcore-dev libx264-dev libfontconfig1-dev libcairo2-dev libgdk-pixbuf2.0-dev libpango1.0-dev libgtk2.0-dev libgtk-3-dev libatlas-base-dev gfortran libhdf5-dev libhdf5-serial-dev libhdf5-103 python3-dev v41-utils || { echo "Package installation failed! Aborting..."; exit 1; }
 echo "Required packages installed."
 
 # Remove the existing app directory if it exists
 echo "Checking for existing installation..."
-if [ -d "/home/pi/BlossomApp" ]; then
+if [ "$(pwd)" == "/home/pi/BlossomApp" ]; then
+    echo "You're in the directory that is about to be removed. Exiting."
+    exit 1
+else
+  if [ -d "/home/pi/BlossomApp" ]; then
     echo "Removing existing installation..."
     rm -rf /home/pi/BlossomApp
+  fi
+  echo "Existing installation removed (if any)."
 fi
-echo "Existing installation removed (if any)."
 
 # Clone the repository
 echo "Cloning the repository..."
